+++ tests/lib/rules/no-invalid-html-attribute.js
@@ -211,6 +211,24 @@ ruleTester.run('no-invalid-html-attribute', rule, {
         })
       `,
     },
+    {
+      code: '<link rel="apple-touch-icon" sizes="60x60" href="apple-touch-icon-60x60.png" />',
+    },
+    {
+      code: '<link rel="apple-touch-icon" sizes="76x76" href="apple-touch-icon-76x76.png" />',
+    },
+    {
+      code: '<link rel="apple-touch-icon" sizes="120x120" href="apple-touch-icon-120x120.png" />',
+    },
+    {
+      code: '<link rel="apple-touch-icon" sizes="152x152" href="apple-touch-icon-152x152.png" />',
+    },
+    {
+      code: '<link rel="apple-touch-icon" sizes="180x180" href="apple-touch-icon-180x180.png" />',
+    },
+    {
+      code: '<link rel="mask-icon" href="/safari-pinned-tab.svg" color="#fff" />',
+    },
   ],
   invalid: [
     {